python/user-auth-app/requirements.txt
Flask==3.1.0
protobuf==5.29.2
google_auth_oauthlib==1.2.1
google-apps-chat==0.2.0
google-cloud-firestore==2.19.0
